Consolidate all the OSC controls into fixtures themselves.
