+++ src/fixture/profile/aquarius.rs
@@ -7,7 +7,7 @@ use strum_macros::{Display as EnumDisplay, EnumIter, EnumString};
 use crate::fixture::prelude::*;
 use crate::osc::prelude::*;
 use crate::util::bipolar_fader_with_detent;
-use crate::{master::FixtureGroupControls, util::bipolar_to_split_range};
+use crate::util::bipolar_to_split_range;
 
 #[derive(Default, Debug)]
 pub struct Aquarius {
