+++ src/fixture/profile/color.rs
@@ -6,8 +6,6 @@ use anyhow::{bail, Result};
 use num_derive::{FromPrimitive, ToPrimitive};
 use number::{Phase, UnipolarFloat};
 
-use crate::master::FixtureGroupControls;
-
 use crate::fixture::prelude::*;
 use crate::osc::prelude::*;
 use strum_macros::{Display as EnumDisplay, EnumIter, EnumString};
