+++ src/fixture/profile/hypnotic.rs
@@ -2,7 +2,8 @@
 
 use number::BipolarFloat;
 
-use crate::fixture::prelude::*;use crate::osc::prelude::*;
+use crate::fixture::prelude::*;
+use crate::osc::prelude::*;
 use crate::util::bipolar_to_split_range;
 use num_derive::{FromPrimitive, ToPrimitive};
 
@@ -127,3 +128,22 @@ impl AnimationTarget {
         false
     }
 }
+
+const GROUP: &str = Hypnotic::NAME.0;
+
+const RED_LASER_ON: Button = button(GROUP, "RedLaserOn");
+const GREEN_LASER_ON: Button = button(GROUP, "GreenLaserOn");
+const BLUE_LASER_ON: Button = button(GROUP, "BlueLaserOn");
+
+impl Hypnotic {
+    pub fn map_controls(map: &mut GroupControlMap<ControlMessage>) {
+        use StateChange::*;
+        RED_LASER_ON.map_state(map, RedLaserOn);
+        GREEN_LASER_ON.map_state(map, GreenLaserOn);
+        BLUE_LASER_ON.map_state(map, BlueLaserOn);
+
+        map.add_bipolar("Rotation", |v| Rotation(bipolar_fader_with_detent(v)));
+    }
+}
+
+impl HandleOscStateChange<StateChange> for Hypnotic {}
