+++ src/fixture/profile/rush_wizard.rs
@@ -4,7 +4,8 @@ use log::error;
 use number::{BipolarFloat, UnipolarFloat};
 
 use super::generic::{GenericStrobe, GenericStrobeStateChange};
-use crate::fixture::prelude::*;use crate::osc::prelude::*;
+use crate::fixture::prelude::*;
+use crate::osc::prelude::*;
 use crate::util::{bipolar_to_range, bipolar_to_split_range, unipolar_to_range};
 use strum_macros::{Display as EnumDisplay, EnumIter, EnumString};
 
@@ -154,6 +155,58 @@ impl Color {
             Red => 176,
             Orange => 179,
             Green => 182,
+        }
+    }
+}
+
+const GROUP: &str = RushWizard::NAME.0;
+const COLOR: &str = "Color";
+
+const GOBO_SELECT: RadioButton = RadioButton {
+    group: GROUP,
+    control: "Gobo",
+    n: 16,
+    x_primary_coordinate: false,
+};
+
+const TWINKLE: Button = button(GROUP, "Twinkle");
+
+impl EnumRadioButton for Color {}
+
+impl RushWizard {
+    pub fn map_controls(map: &mut GroupControlMap<ControlMessage>) {
+        use StateChange::*;
+        map.add_unipolar("Dimmer", Dimmer);
+        map_strobe(map, "Strobe", &wrap_strobe);
+        map.add_enum_handler(COLOR, ignore_payload, |c, _| Color(c));
+        TWINKLE.map_state(map, Twinkle);
+        map.add_unipolar("TwinkleSpeed", TwinkleSpeed);
+        GOBO_SELECT.map(map, Gobo);
+        map.add_bipolar("DrumRotation", |v| {
+            DrumRotation(bipolar_fader_with_detent(v))
+        });
+        map.add_bipolar("DrumSwivel", DrumSwivel);
+        map.add_bipolar("ReflectorRotation", |v| {
+            ReflectorRotation(bipolar_fader_with_detent(v))
+        });
+    }
+}
+
+fn wrap_strobe(sc: GenericStrobeStateChange) -> ControlMessage {
+    StateChange::Strobe(sc)
+}
+
+impl HandleOscStateChange<StateChange> for RushWizard {
+    fn emit_osc_state_change<S>(sc: StateChange, send: &S)
+    where
+        S: crate::osc::EmitOscMessage + ?Sized,
+    {
+        match sc {
+            StateChange::Color(c) => {
+                c.set(GROUP, COLOR, send);
+            }
+            StateChange::Gobo(v) => GOBO_SELECT.set(v, send),
+            _ => (),
         }
     }
 }
