+++ src/master.rs
@@ -7,14 +7,10 @@ use tunnels::clock_server::StaticClockBank;
 
 use crate::animation::AnimationUIState;
 use crate::channel::{ChannelStateEmitter, Channels};
+use crate::fixture::generic::{map_strobe, GenericStrobe, GenericStrobeStateChange};
+use crate::fixture::prelude::*;
 use crate::fixture::Patch;
-use crate::osc::{GroupControlMap, HandleStateChange, OscControlMessage};
-use crate::{
-    fixture::generic::{GenericStrobe, GenericStrobeStateChange},
-    osc::EmitControlMessage,
-};
-
-pub use crate::fixture::FixtureGroupControls;
+use crate::osc::{prelude::*, EmitControlMessage};
 
 pub struct MasterControls {
     strobe: Strobe,
@@ -109,3 +105,24 @@ pub struct Strobe {
     pub state: GenericStrobe,
     pub use_master_rate: bool,
 }
+
+pub(crate) const GROUP: &str = "Master";
+
+const USE_MASTER_STROBE_RATE: Button = button(GROUP, "UseMasterStrobeRate");
+const REFRESH_UI: Button = button(GROUP, "RefreshUI");
+
+impl MasterControls {
+    pub fn map_controls(map: &mut GroupControlMap<ControlMessage>) {
+        map_strobe(map, "Strobe", &wrap_strobe);
+        USE_MASTER_STROBE_RATE.map_state(map, |v| {
+            ControlMessage::State(StateChange::UseMasterStrobeRate(v))
+        });
+        REFRESH_UI.map_trigger(map, || ControlMessage::RefreshUI)
+    }
+}
+
+fn wrap_strobe(sc: GenericStrobeStateChange) -> ControlMessage {
+    ControlMessage::State(StateChange::Strobe(sc))
+}
+
+impl HandleOscStateChange<StateChange> for MasterControls {}
