--- src/osc/profile/mod.rs
@@ -1,16 +1,0 @@
-mod colordynamic;
-mod comet;
-mod dimmer;
-mod faderboard;
-pub mod generic;
-mod h2o;
-mod hypnotic;
-mod lumasphere;
-mod radiance;
-mod rush_wizard;
-mod solar_system;
-mod starlight;
-mod swarmolon;
-mod uv_led_brick;
-mod venus;
-mod wizard_extreme;
